+++ server/Models/TicketModel/TicketCreateDTO.cs
@@ -11,11 +11,11 @@ namespace server.Models.TicketModel
 
         public string Description { get; set; }
 
-        public short TypeLabel { get; set; }
+        public byte TypeLabel { get; set; }
 
-        public short Reproducibility { get; set; }
+        public byte Reproducibility { get; set; }
 
-        public short Severity { get; set; }
+        public byte Severity { get; set; }
 
         public List<string> ImageLinks { get; set; }
     }
